requirements.txt
torch==2.4.1
transformers==4.45.2
datasets==3.0.1
accelerate==1.0.1
deepspeed==0.15.2
packaging
ninja
tqdm
wandb
numpy
Pillow